"macros" interpreter: evaluate conditions
Evaluate "if"/"unless" statements from structured commands.

The placeholder condition, "_", can be used to reference the
value of the previous conditional expression.
(In case of "unless", that is the negated value
of the conditional statement.)

See 26ff0030f352a5f7c7dac1740f8b93d24593bac9 for lang details.

The "hardware-match" condition gets recognized by the parser,
but not by the interpreter, because the underlying feature,
"hardware detection", is not there yet.