add standalone install info file
This file defines the fs layout of the project source directory.

Config files are looked up in
  "~/.config/kernelconfig", "<prjroot>/local/config", "<prjroot>/config"
(in that order, first match)

and include files are looked up in the "include" subdirectories of
the previous paths.

Data files are looked up in
   "<prjroot>/local/data", "<prjroot>/files/data"
(in that order, first match),
but this only of theoretical interest as there are no data files so far.