"script" conf source: temporaraily disable auto vars
For "embedded script" config source types,
the script itself should be formatted.

To do that halfway efficiently, it is necessary to
detect whether the script can be "statically" formatted
(in init_from_*()) or needs to created for each invocation
of get_configuration_basis().
Since that is not done yet, auto vars are disabled for now.