sources, env,fmt vars: properly handle None values
If an env var name is mapped to "None",
it should be removed from the environment when creating a subprocess.
However, the str() conversion eliminated this possibility.
Fixed by passing objects as-is, subproc is responsible for creating
proper env vars (remove "None" vars, str-convert others).

If an fmt var name is mapped to "None",
it should not exist in the fmt vars dict.
Fixed by filtering out "None" vars.
Also added "S" to the fmt vars as required per documentation.