conf sources: strip all values from source def file
The ini parser allows multiline values.

If the value's first line is left empty
and the actual value starts on the second line,
then the read value starts with a newline char:

   option =
     value

Resulting in {"option": "\nvalue"}.

Since whitespace around values is not considered significant,
simply strip all values. For "dict_sets" options,
the (w for w in _.split() if w) dict-comp does that already.