sourcedef: parse "command" field
The "Command=" field in the [source] section of the source def file
can be used to

(a) specify a command (implicit source type: command)

(b) specify the script call (implicit source type: script)

To differentiate the implicit type, the command gets parsed.
If it contains the "script_file" format var,
the "script" type is assumed, and otherwise "command".