add curated source: Ubuntu
This is based on the "ubuntu" sources script from the original project,
but has been converted to a "pym"-type source that allows interfacing
with kernelconfig, and also does a bit more error-checking.
See the comment header in config/sources/ubuntu for details.

Additionally, a few more target archs are supported now, namely armh, arm64.