add "Debian" config source
For Debian's config files,
it is necessary to clone a big git repo
or work around that with git ls-remote and http downloads,

parse debcontrol-like files with [section] headers,

and implement its splitconfig layout:
  (base X) kernelarch X arch X arch flavour X featureset

As of this commit, this is done.

However, proper arch name mappings in the source definition file
are mostly todo.