main script: add "hwdetect-from-file" stub
Add a -H, --hwdetect option that instructions kernelconfig
to enable config options for hardware (and possibly fstypes)
previously detected with the hwcollector script.

-H, --hwdetect disables "hwdetect" instructions in "macros" files,
mixing them is not possible.

With this commit, the main script is able to accept and recognize
hwdetect files, the actual functionality is not implemented yet
(this will mostly be done in kernel/hwdetection/detector),
a NotImplementedError will be raised instead.