pm integration: get pkgs inheriting linux-info.eclass
Also adds a small main script that prints out the build-time
value of the CONFIG_CHECK variable.

Note that this is just preliminary work on pm-integration
and not accessible in config generation code.
It is planned to create a dummy overlay for all packages
inheriting linux-info.eclass, and calculate the correct
CONFIG_CHECK value at config generation time.