pm portagevdb: re-evaluate CONFIG_CHECK from ebuilds
wip - most code paths are not complete yet,
especially error cases which raise exceptions instead of logging

Using ebuild(1), the pkg_pretend()/pkg_setup() phases of ebuilds
inheriting linux-info are re-run with a modified linux-info.eclass
that captures the value of CONFIG_CHECK in a temporary file,
which is then parsed by kernelconfig to get the config option recommendations.

This commit adds the functionality up to reading the config check tmpfile.