add "config choice modules"
Instead of adding the logic for config suggestions (hw, pm)
to the lang interpreter or config choices/config gen directly,
add a uniform way to suggest config options, ChoiceModule objects,
which need to implement a get_suggestions(**kwargs) method.

The modules are organized by a ChoiceModules object,
which also takes care of lazy-loading them.
(For hwdetection, this is unnecessary, for pm it is not.)