"macros" lang interpreter: process "packages" command
Allows to get the package build-time CONFIG_CHECK value
for all installed packages or re-evaluate it by running
the pkg_pretend()/pkg_setup() phases of all ebuilds.

The reported config options are then enabled or disabled, respectively.

Note that processing of the suggested config options is currently *BROKEN*,
packages recommended config options that do not exist, e.g. GRKERNSEC_ options.
(despite providing the correct kernel srctree to the ebuilds,
but that's how CONFIG_CHECK works).
This will be fixed in a follow-up commit.