src/, lkc bindings: get symbol defaults
When resolving a configuration,
as an alternative to making any prompt of a symbol visible,
making one of its defaults visible is also an option.

This is necessary necessary for some symbols
that do have a conditional prompt, for example HPET_TIMER:

   config HPET_TIMER
      def_bool X86_64
      prompt ___ if X86_32

where kernelconfig tries to enable X86_32 for a X86_64 config
to make HPET_TIMER visible (and errors out)
while it should just pick the default.

This commit adds the "C"-side support for dealing with symbol
defaults, a SymbolViewObject method that returns defaults
as 2-tuple (dir_dep, vis_dep).