kconfig symbol abc: split base class
This is a preparation step for symbol defaults, which are,
essentially, objects with a dir_dep and a vis_dep expression,
and thus similar to kconfig symbol objects.

Kconfig symbol classes lose "collections.abc.Hashable" as
parent class, this shouldn't have any impact on the code.

(isinstance(_, Hashable) will no longer work,
but it isn't used in the code.)