conf sources: fix: empty source section in settings
If no configuration source is requested in the [source] section
of the settings file, <srctree>/.config should be used.

However, the existing code is relict back from when
get_configuration_source_from_settings() was part of
get_configuration_basis_from_settings(),
and a configuration basis was a single file (now: list of files).

This commit fixes that:

* get_configuration_source_from_settings() returns (True, None)
  if no source configured in settings

* get_configuration_basis_from_settings() creates the appropriate
  configuration basis for that case