kernel source info: add "real kernelversion"
For configuration sources, it can be useful to pretend
a different kernel version.
This commit adds the notion of a "real" kernel version
"<obj>.real_kernelversion" to KernelInfo objects,
which is the same as "<obj>.kernelversion".

The idea is to pass a proxy object as source_info
to objects that should receive a fake kver.
"<obj>.real_kernelversion" will then contain the real version,
and "<obj>.kernelversion" the fake version.